sauvegarde des modifications locale

+++ src/main/java/org/zzn/hospital/Repositories/OrdonnanceMedecineRepository.java
@@ -3,5 +3,5 @@ package org.zzn.hospital.Repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.zzn.hospital.Models.OrdonnanceMedecine;
 
-public interface OrdonnanceMedecineRepository extends JpaRepository<OrdonnanceMedecine, Integer> {
+public interface OrdonnanceMedecineRepository extends JpaRepository<OrdonnanceMedecine, Long> {
 }
